QRvisu/app/requirements.txt
fastapi
uvicorn
pyzbar
pillow
python-multipart
# opcional: opencv-python
